+++ src/models/instances/customItem.ts
@@ -30,12 +30,12 @@ export default function defineCustomItems(sequelize: Sequelize) {
         defaultValue: false,
       },
       quantity: {
-        type: DataTypes.FLOAT(10, 2),
+        type: DataTypes.DECIMAL(10, 2),
         allowNull: false,
         defaultValue: 0
       },
       price: {
-        type: DataTypes.FLOAT(10, 2),
+        type: DataTypes.DECIMAL(10, 2),
         allowNull: false,
         defaultValue: 0
       },
